+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * [HexaPDF::Layout::ListBox] to hide marker in case of splitting list items with
   multiple boxes
 * [HexaPDF::Layout::ListBox] to create independent marker boxes for all markers
+* [HexaPDF::Layout::ListBox] to correctly respect a set height
 
 
 ## 0.39.1 - 2024-03-20